+++ package.json
@@ -7,6 +7,11 @@
     "@testing-library/jest-dom": "4.2.4",
     "@testing-library/react": "9.3.2",
     "@testing-library/user-event": "7.1.2",
+    "@types/jest": "26.0.15",
+    "@types/node": "12.0.0",
+    "@types/react": "16.9.53",
+    "@types/react-dom": "16.9.8",
+    "@types/react-redux": "^7.1.15",
     "eslint-plugin-prettier": "2.7.0",
     "eslint-plugin-react": "7.18.0",
     "node-sass": "4.14.1",
@@ -16,7 +21,8 @@
     "react-redux": "7.1.3",
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.1"
+    "react-scripts": "4.0.1",
+    "typescript": "4.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -25,7 +31,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
